requirements.txt
langchain_community
langchain_openai
requests
fastapi
uvicorn
streamlit
python-dotenv
pandas
langchain
openai
faiss-cpu
